Moved fiels and modified endpoints
