+++ nextjs-blog/.gitignore
@@ -28,3 +28,7 @@ yarn-error.log*
 .env.development.local
 .env.test.local
 .env.production.local
+
+# package locks
+package.json
+package-lock.json
