+++ nextjs-blog/pages/api/client/updateClient.js
@@ -1,4 +1,5 @@
-import clientPromise from "../../lib/mongodb";
+// import clientPromise from "../../lib/mongodb";
+import clientPromise from "../../../lib/mongodb";
 
 export default async (req, res) => {
   try {
